requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
defusedxml==0.8.0rc2
dj-rest-auth==7.0.0
Django==5.1.3
django-allauth==65.3.1
django-cors-headers==4.6.0
django-filter==24.3
django-jazzmin==3.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1*
djoser==2.3.1
drf-spectacular==0.28.0
frozenlist==1.5.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
multidict==6.1.0
oauthlib==3.2.2
pillow==11.0.0
propcache==0.2.1
psycopg==3.2.3
pycparser==2.22
PyJWT==2.10.1
python3-openid==3.2.0
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.22.3
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.2
twilio==9.3.7
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
yarl==1.18.3
